+++ cpp/src/io/writer_public.cpp
@@ -68,11 +68,15 @@ Node Writer::AddNode(Page &page, const VoxelKey &key, las::Points &points)
 {
     auto header = file_->GetLasHeader();
     if (points.PointFormatID() != header.point_format_id || points.PointRecordLength() != header.point_record_length)
-        throw std::runtime_error("New points must be of same format and size.");
+        throw std::runtime_error("Writer::AddNode: New points must be of same format and size.");
 
-    // Check that points fit within the key
-    if (!points.Within(Box(key, GetLasHeader())))
-        throw std::runtime_error("Points must be within the key");
+    // Check that all points are within the file bounds
+    if (!points.Within(Box(header.min, header.max)))
+        std::cout << "Writer::AddNode Warning: Adding points outside of las header bounds!" << std::endl;
+
+    // Check that all points are within the node
+    if (!points.Within(Box(key, header)))
+        throw std::runtime_error("Writer::AddNode: Points must be within the key");
 
     std::vector<char> uncompressed = points.Pack();
     return AddNode(page, key, uncompressed);